refactor(subagent): move the pause-checkpoint writer to its own module
`scripts/ci/check-openhuman-rust-layout.mjs` pins `runner.rs` at 1769 lines
and `progress_bridge.rs` at 1547 as legacy exceptions, and the gate is
deliberately monotonic — a pinned file may not grow. Adding
`write_pause_checkpoint` pushed runner.rs to 1801 and the ledger comment
pushed progress_bridge.rs to 1550, so both tripped it.

The writer moves to `ops/pause_checkpoint.rs` (named to distinguish it from
the existing `ops/checkpoint.rs`, which summarises a cap-hit run rather than
persisting a pause), taking its tests with it. `progress_bridge.rs` drops the
now-redundant local by asking `checkpoint_path` directly.

No behaviour change.
